feat: implement CreateVehicleMake

+++ src/Domain/Identity/Entities/UserRole.cs
@@ -2,8 +2,8 @@
 {
     public partial class UserRole
     {
-        public int UserId { get; set; }
-        public int RoleId { get; set; }
+        public long UserId { get; set; }
+        public short RoleId { get; set; }
         public DateTime AssignedAtUtc { get; set; }
         public virtual User User { get; set; } = null!;
         public virtual Role Role { get; set; } = null!;
